[front] rename folder `game/localApi` to `game/utils`
