+++ front/app/src/game/utils/GameWorker.js
