+++ front/app/src/game/utils/calculateNextAiPosition.js
