+++ front/app/src/game/utils/characters.js
