+++ front/app/src/game/view-game-offline.ce.js
@@ -1,7 +1,7 @@
 import './components/game-select-player.ce.js';
 import './components/game-play.ce.js';
 import { redirectTo } from '@/router.js';
-import { characters } from './localApi/characters.js';
+import { characters } from './utils/characters.js';
 import { isAuthenticated, getProfile } from '@/auth.js';
 
 class ViewGameOffline extends HTMLElement {
